+++ pyproject.toml
@@ -5,7 +5,7 @@ build-backend = "setuptools.build_meta"
 [project]
 name = "gnn-benchmark"
 version = "0.1.0"
-dependencies = ["pandas", "numpy"]
+dependencies = ["pandas", "numpy", "gdown"]
 
 [tool.setuptools]
 packages = [
